pyproject.toml
[project]
name = "website-quality-agent"
version = "0.1.0"
description = "Automated website quality analyzer - crawls sites, detects issues, generates reports"
readme = "PLAN.md"
requires-python = ">=3.11"
dependencies = [
    # Web crawling
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "playwright>=1.40.0",

    # LLM for content analysis
    "openai>=1.0.0",

    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Database
    "aiosqlite>=0.19.0",

    # PDF generation
    "weasyprint>=60.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
website-agent = "website_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/website_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
